issue/14-b Moved view into separate file
